VyaparIQ-Medical/requirements.txt
# Core Dependencies
streamlit==1.31.0
boto3==1.34.34
python-dotenv==1.0.0

# AWS SDK
botocore==1.34.34

# Data Processing
pandas==2.2.0
numpy==1.26.3
Pillow==10.2.0

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-mock==3.12.0
moto==5.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0